deduplicate win32-gl bridges
